+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 ### Removed
 
 ### Fixed
+- Fix swapped column names "Locals" and "Stack" in method frames dialog (GitHub issue #9)
+
 ## [2.0.2] - 2021-07-28
 ### Fixed
 - Fix compatibility with IntelliJ 2021.2